refactor existing tests for reusing parts
