+++ tests/components/foscam/test_config_flow.py
@@ -2,79 +2,15 @@
 
 from unittest.mock import patch
 
-from libpyfoscam.foscam import (
-    ERROR_FOSCAM_AUTH,
-    ERROR_FOSCAM_CMD,
-    ERROR_FOSCAM_UNAVAILABLE,
-    ERROR_FOSCAM_UNKNOWN,
-)
-
 from homeassistant import config_entries
 from homeassistant.components.foscam import config_flow
 from homeassistant.core import HomeAssistant
 from homeassistant.data_entry_flow import FlowResultType
 
+from .conftest import setup_mock_foscam_camera
+from .const import CAMERA_NAME, INVALID_RESPONSE_CONFIG, VALID_CONFIG
+
 from tests.common import MockConfigEntry
-
-VALID_CONFIG = {
-    config_flow.CONF_HOST: "10.0.0.2",
-    config_flow.CONF_PORT: 88,
-    config_flow.CONF_USERNAME: "admin",
-    config_flow.CONF_PASSWORD: "1234",
-    config_flow.CONF_STREAM: "Main",
-    config_flow.CONF_RTSP_PORT: 554,
-}
-OPERATOR_CONFIG = {
-    config_flow.CONF_USERNAME: "operator",
-}
-INVALID_RESPONSE_CONFIG = {
-    config_flow.CONF_USERNAME: "interr",
-}
-CAMERA_NAME = "Mocked Foscam Camera"
-CAMERA_MAC = "C0:C1:D0:F4:B4:D4"
-
-
-def setup_mock_foscam_camera(mock_foscam_camera):
-    """Mock FoscamCamera simulating behaviour using a base valid config."""
-
-    def configure_mock_on_init(host, port, user, passwd, verbose=False):
-        product_all_info_rc = 0
-        dev_info_rc = 0
-        dev_info_data = {}
-
-        if (
-            host != VALID_CONFIG[config_flow.CONF_HOST]
-            or port != VALID_CONFIG[config_flow.CONF_PORT]
-        ):
-            product_all_info_rc = dev_info_rc = ERROR_FOSCAM_UNAVAILABLE
-
-        elif (
-            user
-            not in [
-                VALID_CONFIG[config_flow.CONF_USERNAME],
-                OPERATOR_CONFIG[config_flow.CONF_USERNAME],
-                INVALID_RESPONSE_CONFIG[config_flow.CONF_USERNAME],
-            ]
-            or passwd != VALID_CONFIG[config_flow.CONF_PASSWORD]
-        ):
-            product_all_info_rc = dev_info_rc = ERROR_FOSCAM_AUTH
-
-        elif user == INVALID_RESPONSE_CONFIG[config_flow.CONF_USERNAME]:
-            product_all_info_rc = dev_info_rc = ERROR_FOSCAM_UNKNOWN
-
-        elif user == OPERATOR_CONFIG[config_flow.CONF_USERNAME]:
-            dev_info_rc = ERROR_FOSCAM_CMD
-
-        else:
-            dev_info_data["devName"] = CAMERA_NAME
-            dev_info_data["mac"] = CAMERA_MAC
-
-        mock_foscam_camera.get_product_all_info.return_value = (product_all_info_rc, {})
-        mock_foscam_camera.get_dev_info.return_value = (dev_info_rc, dev_info_data)
-
-        return mock_foscam_camera
-
-    mock_foscam_camera.side_effect = configure_mock_on_init
 
 
 async def test_user_valid(hass: HomeAssistant) -> None:
